requirements.txt
fastapi[standard]
firebase-admin
google-cloud-storage
google-cloud-aiplatform